move cephstore to a separate package
