+++ pkg/scsi/backingstore/cephstore/cephstore_linux.go
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package backingstore
+package cephstore
 
 import (
 	"fmt"
